requirements.txt
opencv-python
pyautogui
python-dotenv
inference-sdk
numpy